Moved Drawer to its own file
Settings initialization were a PITA!
